requirements.txt
asttokens==2.4.1
certifi==2022.12.7
charset-normalizer==2.1.1
cmake==3.25.0
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.9.0
fonttools==4.46.0
idna==3.4
imageio==2.33.1
importlib-resources==6.1.1
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
lit==15.0.7
lmdb==1.4.1
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mpmath==1.3.0
networkx==3.0
numpy==1.24.1
opencv-python==4.8.1.78
packaging==23.2
parso==0.8.3
pexpect==4.9.0
Pillow==9.3.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
requests==2.28.1
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
stack-data==0.6.3
sympy==1.12
threadpoolctl==3.2.0
tifffile==2023.12.9
torch==2.0.0+cu118
torchaudio==2.0.1+cu118
torchvision==0.15.1+cu118
tqdm==4.66.1
traitlets==5.14.0
triton==2.0.0
typing_extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.12
zipp==3.17.0
